# protocol/bc: use pointers for func receivers