The AssetAmount and Outpoint WriteTo functions were previously declared
on raw structs. Each call to these functions would cause an allocation.
In a recent run of benchcore these two function calls collectively
allocated 4.5% of all objects.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkAssetAmountWriteTo-8     65.5          38.8          -40.76%
BenchmarkOutpointWriteTo-8        69.0          37.4          -45.80%

benchmark                         old allocs     new allocs     delta
BenchmarkAssetAmountWriteTo-8     1              0              -100.00%
BenchmarkOutpointWriteTo-8        1              0              -100.00%

benchmark                         old bytes     new bytes     delta
BenchmarkAssetAmountWriteTo-8     48            0             -100.00%
BenchmarkOutpointWriteTo-8        48            0             -100.00%
```